+++ inst/shiny/ui.R
@@ -56,7 +56,7 @@ shinyUI(fluidPage(
           uiOutput("RawValues"),
           selectInput("Scale", label = "Norm Scale", choices = c("T", "IQ", "z"), selected = "T"),
           actionButton(inputId = "DoDataPreparation", label = "Prepare Data"),
-
+          downloadButton("downloadData", "Download Data"),
           tags$br(),
           tags$br(),
           # Additional options (explanatory variable with default grouping, ranking method, number of powers)
@@ -100,11 +100,14 @@ shinyUI(fluidPage(
           tags$p("Here, you can calculate a regression model that models the original data as close as possible, while smoothing the curves and eliminating noise. After hitting the button, the regression function for a possible model is shown. The plot displays the information criteria for the different models, beginning with the model with one terms up to the maximum. The model should have a high R2 with as few terms as possible. The information of the plot is again displayed as a table below the chart. To display plots of observed vs. fitted raw and norm scores, please use the plotting options on the next tab. There, you can check the percentile curves as well and inspect the curves of the different plausible models."),
           tags$br(), tags$b("HINT: Please ensure that the data is loaded and prepared, before starting the modeling. In case of k > 4, the calculation will take a few seconds."),
           tags$br(), tags$br(), tags$b("HINT: Look out for an 'elbow' in the chart and try that number of terms for your model. Afterwards, plot the percentiles to check the curves."),
-          tags$br(), tags$br(), actionButton(
-            inputId = "CalcBestModel",
+          tags$br(), tags$br(),
 
+          actionButton(
+            inputId = "CalcBestModel",
             label = "Model Data"
           ),
+
+          downloadButton("downloadModel", "Download Model"),
 
           tags$br(),
           # Additional options (R^2, terms, type for printSubset)
@@ -125,9 +128,11 @@ shinyUI(fluidPage(
           verbatimTextOutput("BestModel6"),
           verbatimTextOutput("BestModel7"),
           tags$br(),
-          tags$br(),
           # tags$h4("Information Function, Subset Specifics and Fitted Values"),
           # tags$p("The plot shows the informationcriteria for the different models, beginning with the model with one terms up to the maximum. The model should have a high R2 with as few terms as possible. The information of the plot is again displayed as a table below the chart. On the bottom of the page, you can see, how well the observed data are fitted by the model."),
+          # plotOutput("PlotPercentiles", width = "100%", height = "600px"),
+          withSpinner(plotOutput("modelPlot", width = "100%", height = "600px"), type = 5),
+          tags$br(),
           withSpinner(plotOutput("PlotWL", width = "100%", height = "600px"), type = 5),
           tags$br(),
           withSpinner(dataTableOutput("PrintSubset"), type = 5),